Move expedition receive handler to kca
